==> prgs/20_Descriptives_vs.Rmd <==
---
title: Klasse im Puls
author: 
  - Christoph Wunder
date: "`r format(Sys.time(), '%B %d, %Y')`"
output:
  html_document:
    fig_caption: yes
    theme: spacelab 
    highlight: pygments
    toc: TRUE
    toc_depth: 2
    number_sections: TRUE
    toc_float:
      smooth_scroll: FALSE
bibliography: literature.bib
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(comment=NA)
rm(list=ls())
```
```{r, include=TRUE, message = FALSE}
library(tidyverse)
library(Hmisc)
library(mosaic)
library(foreign)
library(lme4)
library(rstanarm)
library(brms)
library(tableone)
library(RColorBrewer)
library(xtable)
library(knitr)
library(kableExtra)
library(ggplot2)
library(extrafont)
library(grid)
library(gridExtra)
# windowsFonts(CMU = windowsFont("CMU Serif")) # Define CMU font as the computer modern serif font that is used in my latex template

theme_set(theme_bw(base_family = "CMU Serif") +
            theme(text = element_text(size = 16)))
#           text=element_text(size = 20)) # Set the gg-theme to use the latex font

# Sys.setenv(LANG="en") # get better error messages for googling
```


# Data

```{r}
welle1 <- readRDS("../data/welle1.RDS")
welle2 <- readRDS("../data/welle2.RDS")
welle3 <- readRDS("../data/welle3.RDS")
welle4 <- readRDS("../data/welle4.RDS")
welle5 <- readRDS("../data/welle5.RDS")
```


## Pre-treatment Analysis

```{r}

sum_satisfaction <-
  tibble(
    variable = c("Life satisfaction",
                "Satisfaction friends",
                "Satisfaction class",
                "Satisfaction school"),
    mean = c(mean( welle1$lsat, na.rm=T),
                  mean(welle1$sat_friends, na.rm=T),
                  mean(welle1$sat_class, na.rm=T),
                  mean(welle1$sat_school, na.rm=T)),
    sd = c(sd(welle1$lsat, na.rm=T),
           sd(welle1$sat_friends, na.rm=T),
           sd(welle1$sat_class, na.rm=T),
           sd(welle1$sat_school, na.rm=T))
    )
    
saveRDS(sum_satisfaction, file = "../results/sum_satisfaction.RDS")

```

```{r}
lsat_bar <- ggplot(welle1, aes(x=lsat)) +
  geom_bar((aes(y=(..count..)/sum(..count..))), fill = "#004c99") +
  labs(x="Life satisfaction", y="Density") +
  scale_x_discrete( limits = seq(0, 10))
ggsave("../figures/lsat_bar.png", lsat_bar, width = 7, height = 5, units = "in")
```



```{r}
# Mean differences life satisfaction treatment group vs. control group
su_welle1 <- welle1 %>% group_by(school, treated) %>%
                      summarise( mean.lsat = mean(lsat, na.rm = TRUE), 
                                 sd.lsat=sd(lsat, na.rm=TRUE) )
lsat_pre_treat <- ggplot( su_welle1, 
                          aes(x=school, y=mean.lsat, fill=treated) ) +
  geom_bar ( stat="identity", position = "dodge" ) + 
  coord_cartesian( ylim = c(0, 11) ) + 
  labs(x = "School", y = "Mean Life Satisfaction",
       title = "Life satisfaction", 
       caption = "Mean pretreatment life satisfaction (+ 1 sd) (wave 1) for treatment and control group per school") +
  geom_errorbar( aes(ymin=mean.lsat-sd.lsat, ymax=mean.lsat+sd.lsat), 
                 width=.2, position=position_dodge(.9) ) +
  scale_fill_brewer(palette="Blues")

ggsave("../figures/lsat_pre_treat.png", lsat_pre_treat, width = 7, height = 5, units = "in")

lsat_pre_treat
```
```{r}
# Mean differences satisfaction with friends treatment group vs. control group
su_welle1 <- welle1 %>% group_by(school, treated) %>%
                      summarise( mean.sat.friends = mean(sat_friends, na.rm = TRUE), 
                                 sd.sat.friends=sd(sat_friends, na.rm=TRUE) )
sat_friends_pre_treat <- ggplot( su_welle1, 
                          aes(x=school, y=mean.sat.friends, fill=treated) ) +
  geom_bar ( stat="identity", position = "dodge" ) + 
  coord_cartesian( ylim = c(0, 11) ) + 
  labs(x = "School", y = "Mean Satisfaction with Friends",
       title = "Satisfaction with Friends", 
       caption = "Mean pretreatment satisfaction with friends (+ 1 sd) (wave 1) for treatment and control group per school") +
  geom_errorbar( aes(ymin=mean.sat.friends-sd.sat.friends, ymax=mean.sat.friends+sd.sat.friends), 
                 width=.2, position=position_dodge(.9) ) +
  scale_fill_brewer(palette="Blues")

ggsave("../figures/sat_friends_pre_treat.png", sat_friends_pre_treat, width = 7, height = 5, units = "in")

sat_friends_pre_treat
```

```{r}
# Mean differences satisfaction with class treatment group vs. control group
su_welle1 <- welle1 %>% group_by(school, treated) %>%
                      summarise(mean.sat.class = mean(sat_class, na.rm = TRUE), 
                                 sd.sat.class=sd(sat_class, na.rm=TRUE) )
sat_class_pre_treat <- ggplot( su_welle1, 
                          aes(x=school, y=mean.sat.class, fill=treated) ) +
  geom_bar ( stat="identity", position = "dodge" ) + 
  coord_cartesian( ylim = c(0, 11) ) + 
  labs(x = "School", y = "Mean Satisfaction with Class",
       title = "Satisfaction with class", 
       caption = "Mean pretreatment satisfaction with class (+ 1 sd) (wave 1) for treatment and control group per school") +
  geom_errorbar( aes(ymin=mean.sat.class-sd.sat.class, ymax=mean.sat.class+sd.sat.class), 
                 width=.2, position=position_dodge(.9) ) +
  scale_fill_brewer(palette="Blues")

ggsave("../figures/sat_class_pre_treat.png", sat_class_pre_treat, width = 7, height = 5, units = "in")

sat_class_pre_treat
```

```{r}
# Mean differences satisfaction with school treatment group vs. control group
su_welle1 <- welle1 %>% group_by(school, treated) %>%
                      summarise( mean.sat.school = mean(sat_school, na.rm = TRUE), 
                                 sd.sat.school=sd(sat_school, na.rm=TRUE) )
sat_school_pre_treat <- ggplot( su_welle1, 
                          aes(x=school, y=mean.sat.school, fill=treated) ) +
  geom_bar ( stat="identity", position = "dodge" ) + 
  coord_cartesian( ylim = c(0, 11) ) + 
  labs(x = "School", y = "Mean Satisfaction with School",
       title = "Satisfaction with school", 
       caption = "Mean pretreatment satisfaction with school (+ 1 sd) (wave 1) for treatment and control group per school") +
  geom_errorbar( aes(ymin=mean.sat.school-sd.sat.school, ymax=mean.sat.school+sd.sat.school), 
                 width=.2, position=position_dodge(.9) ) +
  scale_fill_brewer(palette="Blues")

ggsave("../figures/sat_school_pre_treat.png", sat_school_pre_treat, width = 7, height = 5, units = "in")

sat_school_pre_treat
```
```{r}
grid.arrange(lsat_pre_treat, sat_friends_pre_treat, sat_class_pre_treat, sat_school_pre_treat)
```



```{r}
vars <- c("lsat","female", "sat_friends", "sat_class", "class.prob", "class.excl", "note.deu", "note.mat", "note.mus", "hob.mus.making", "dauer.mus.making", "play.instr", "mus.active", "migback")

names(welle1)[names(welle1) == 'lsat'] <- 'Life satisfaction'
names(welle1)[names(welle1) == 'female'] <- 'Female'
names(welle1)[names(welle1) == 'sat_friends'] <- 'Satisfaction friends'
names(welle1)[names(welle1) == 'sat_class'] <- 'Satisfaction class'
names(welle1)[names(welle1) == 'class.prob'] <- 'Problems in class'
names(welle1)[names(welle1) == 'class.excl'] <- 'Exclusion in class'
names(welle1)[names(welle1) == 'note.deu'] <- 'Grade german'
names(welle1)[names(welle1) == 'note.mat'] <- 'Grade math'
names(welle1)[names(welle1) == 'note.mus'] <- 'Grade music'
names(welle1)[names(welle1) == 'hob.mus.making'] <- 'Hobby music'
names(welle1)[names(welle1) == 'dauer.mus.making'] <- 'Duration music'
names(welle1)[names(welle1) == 'play.instr'] <- 'Instrument'
names(welle1)[names(welle1) == 'mus.active'] <- 'Musically active'
names(welle1)[names(welle1) == 'migback'] <- 'Migration background'

vars <- c("Life satisfaction","Female", "Satisfaction friends", "Satisfaction class", "Problems in class", "Exclusion in class", "Grade german", "Grade math", "Grade music", "Hobby music", "Duration music", "Instrument", "Musically active", "Migration background")

smd <- CreateTableOne(vars = vars, strata = "treated", data = welle1, test = FALSE, includeNA = F, smd = TRUE)

# rownames(smd) <- c("N", "Life satisfaction", "Sex", "male", "female", "unknown", "Satisfaction with friends", "Satisfaction with class", "Grade german", "Grade maths", "Grade music", "Hobby music 'yes' (%)", "Duration music listening", "Duration music making", "not at all", "ca 30 min", "ca 1-2 hours", "ca 3-4 hours", "more than 4 hours", "Playing instrument 'yes' (%)", "Musically active", "Migration background", "  native", "  immigrant", "  unknown" )
# colnames(smdtab) <- c("control group", "treatment group", "SMD")

extract_smd <- ExtractSmd(smd)

saveRDS(object = extract_smd, file = "../results/extract_smd.Rds")


# 2020-09-03 vs: Preferred method to compute smd was "stddiff" but an error occured that I could not fix:
# stddiff.numeric(data=welle1, gcol=4, vcol=5)
# Queston: How can I integrate the resultsing smd in KIP_paper.Rmd?

#ptd <- print(smd, printToggle = FALSE, noSpaces = TRUE)
#xtable(ptd)
```

```{r}
d <- readRDS("../data/d_sat.RDS")

lsat_vs_time_by_school_treated <- ggplot(
  data = subset(d, treated == 1),
  aes(y = lsat, x = welle, group = school)) +
  scale_color_manual(
    values =
      c("1" = "#cce5ff",
        "2" = "#99cbff",
        "3" = "#4ca5ff",
        "4" = "#198bff",
        "5" = "#0065cc",
        "6" = "#004c99")
  ) +
  geom_smooth(aes(color=school), method = "lm", se = F, size = .4, linetype = "twodash") +
  geom_smooth(aes(group = 1), method = "lm", se = F, colour = "#9400d3") +
  labs(colour = "School", x = "Wave", y = "Mean life satisfaction")

ggsave("../figures/lsat_vs_time_by_school_treated.png", lsat_vs_time_by_school_treated, width = 7, height = 5, units = "in")
```

```{r}
d_sum <- welle1 %>% 
  select(pid, welle, migback, female)

genderT <- table(d$female)
genderT

summary(d$female, factor(d$pid))

df <- unique(d[c("pid", "female", "migback")])
table(df$female)
table(df$migback)
as.data.frame(table(df$pid))

nrow(subset(welle1, school ==1))
nrow(subset(welle1, school ==2))
nrow(subset(welle2, school ==3))
nrow(subset(welle2, school ==4))
nrow(subset(welle1, school ==5))
nrow(subset(welle1, school ==6))

n_obs <- tibble(School = c("1", "2", "3", "4", "5", "6", "Sum"),
       "Wave 1" = c(nrow(subset(welle1, school ==1)),
                  nrow(subset(welle1, school ==2)),
                  nrow(subset(welle1, school ==3)),
                  nrow(subset(welle1, school ==4)),
                  nrow(subset(welle1, school ==5)),
                  nrow(subset(welle1, school ==6)),
                  nrow(welle1)
                 ),
       "Wave 2" = c(nrow(subset(welle2, school ==1)),
                  nrow(subset(welle2, school ==2)),
                  nrow(subset(welle2, school ==3)),
                  nrow(subset(welle2, school ==4)),
                  nrow(subset(welle2, school ==5)),
                  nrow(subset(welle2, school ==6)),
                  nrow(welle2)
                 ),
       "Wave 3" = c(nrow(subset(welle3, school ==1)),
                  nrow(subset(welle3, school ==2)),
                  nrow(subset(welle3, school ==3)),
                  nrow(subset(welle3, school ==4)),
                  nrow(subset(welle3, school ==5)),
                  nrow(subset(welle3, school ==6)),
                  nrow(welle3)
                 ),
       "Wave 4" = c(nrow(subset(welle4, school ==1)),
                  nrow(subset(welle4, school ==2)),
                  nrow(subset(welle4, school ==3)),
                  nrow(subset(welle4, school ==4)),
                  nrow(subset(welle4, school ==5)),
                  nrow(subset(welle4, school ==6)),
                  nrow(welle4)
                 ),
       "Wave 5" = c(nrow(subset(welle5, school ==1)),
                  nrow(subset(welle5, school ==2)),
                  nrow(subset(welle5, school ==3)),
                  nrow(subset(welle5, school ==4)),
                  nrow(subset(welle5, school ==5)),
                  nrow(subset(welle5, school ==6)),
                  nrow(welle5)
                 ),
       Sum = c(nrow(subset(d, school ==1)),
               nrow(subset(d, school ==2)),
               nrow(subset(d, school ==3)),
               nrow(subset(d, school ==4)),
               nrow(subset(d, school ==5)),
               nrow(subset(d, school ==6)),
               nrow(d)
              )
        )


n_obs_treat <- tibble(
  " " = c("Treatment group", "Control group"),
  "Wave 1" = c(nrow(subset(welle1, treated==1)),
               nrow(subset(welle1, treated==0))),
  "Wave 2" = c(nrow(subset(welle2, treated==1)),
               nrow(subset(welle2, treated==0))),
  "Wave 3" = c(nrow(subset(welle3, treated==1)),
               nrow(subset(welle3, treated==0))),
  "Wave 4" = c(nrow(subset(welle4, treated==1)),
               nrow(subset(welle4, treated==0))),
  "Wave 5" = c(nrow(subset(welle5, treated==1)),
               nrow(subset(welle5, treated==0)))
)

saveRDS(n_obs, "../results/n_obs.RDS")
kable(n_obs)

saveRDS(n_obs_treat, "../results/n_obs_treat.RDS")
kable(n_obs_treat)
```



# Original computing environment

```{r, results='hide'}
devtools::session_info()
```
