==> prgs/60_Prior_Distribution.Rmd <==
---
title: "Prior predictive distribution"
author: "VS"
date: "24 9 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(comment=NA)
rm(list=ls())
library(tidyverse)
library(brms)
library(sjlabelled)
library(parallel)
library(gridExtra)
library(haven)
library(RColorBrewer)
library(ggmcmc)
library(mcmcplots)
library(cowplot) # cowplot helps align multiple ggplots in a grid for comparison
# options(mc.cores = detectCores()) #detect number of cores for Rstan
# library(extrafont)
# windowsFonts(CMU = windowsFont("CMU Serif")) # Define CMU font as the computer modern serif font that is used in my latex template
# theme_set(theme_bw(base_family = "CMU")) # Set the gg-theme to use the latex font
# Sys.setenv(LANG="en")
library(patchwork)
library(extrafont)
theme_set(theme_bw(base_family = "CMU Serif") +
            theme(text = element_text(size = 16)))# Set the gg-theme to use the latex font
```

```{r}
d <- readRDS("../data/d_sat.RDS")
```

```{r}
test <-
  tibble(lsat = seq(from = 0, to = 10, length.out = 1000)) %>% 
  tidyr::expand(lsat, row = c("flat", "peaked")) %>% 
  arrange(row, lsat) %>% 
  mutate(prior = ifelse(row == "flat", dunif(lsat, -100, 100), dnorm(lsat, 7, 1)),
         likelihood = dnorm(lsat, 7, 2))  %>% 
  group_by(row) %>% 
  mutate(posterior = prior * likelihood / sum(prior * likelihood)) %>% 
  pivot_longer(prior:posterior) %>% 
  ungroup() %>% 
  mutate(name = factor(name, levels = c("prior", "likelihood", "posterior")),
         row  = factor(row, levels = c("flat", "peaked")))

flat <-
  test %>%
  filter(row == "flat") %>% 
  ggplot(aes(x = lsat, y = value)) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  scale_y_continuous(breaks = NULL, limits = NULL) +
  facet_wrap(~name, scales = "free_y") +
  labs(x="", y="") +
  theme(text = element_text(size=14))
flat

peaked <-
  test %>%
  filter(row == "peaked") %>% 
  ggplot(aes(x = lsat, y = value)) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  scale_y_continuous(breaks = NULL, limits = NULL) +
  facet_wrap(~name, scales = "free_y") +
  labs(x="", y="") +
  theme(text = element_text(size=14))
flat_peaked <- flat / peaked
ggsave("../figures/flat_peaked.png", flat_peaked, width = 6, height = 3, units = "in")
flat_peaked

p_intercept_sim <-
  tibble(x = seq(from = -10, to = 10, by = 0.1)) %>%
  
  ggplot(aes(x = x, y = dnorm(x, mean = 7, sd = 2))) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  labs(title = "Intercept ~ dnorm(7,2)", x="", y="")
p_intercept_sim
```
```{r}
get_prior(lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid), data = d)
```


```{r}
p1 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(8, 2), class = Intercept),
                     prior(normal(0, 10), class = b),
                     prior(cauchy(0, 1), class = sd),
                     prior(exponential(0.5), class = sigma)),
                     file = "../results/prior_p1")

p2 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(8, 2), class = Intercept),
                     prior(normal(0, 5), class = b),
                     prior(cauchy(0, 1), class = sd),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_p2")

p3 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(8, 2), class = Intercept),
                     prior(normal(0, 1), class = b),
                     prior(cauchy(0, 1), class = sd),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_p3")

p1.1 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(8, 2), class = Intercept),
                     prior(normal(0, 1), class = b),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_p1.1")
```

```{r}
p1_predict <- posterior_epred(object = p1, nsamples = 100)
p1_long <- gather(data.frame(p1_predict))
p1_long
p1_long$prior <- rep("3",142900)

p1.1_predict <- posterior_epred(object = p1.1, nsamples = 100)
p1.1_long <- gather(data.frame(p1.1_predict))
p1.1_long
p1.1_long$prior <- rep("3",142900)
```
```{r}
p2_predict <- posterior_epred(object = p2, nsamples = 100)
p2_long <- gather(data.frame(p2_predict))
p2_long
p2_long$prior <- rep("2",142900)
```
```{r}
p3_predict <- posterior_epred(object = p3, nsamples = 100)
p3_long <- gather(data.frame(p3_predict))
p3_long
p3_long$prior <- rep("1",142900)
```

```{r}
all <- bind_rows(p3_long, p2_long, p1_long)

prior_predict <- ggplot(all, aes(x=value, y=..density..)) +
  geom_histogram(bins = 200, fill = "#004c99", alpha = .8) +
  xlab("Predicted life satisfaction") + 
  ylab("Density") +
  xlim(-30,30) +
  facet_wrap(vars(prior), labeller = labeller(
  prior = c("1" = "N(0,1)", "2" = "N(0,5)", "3" = "N(0,10)"))) +
  geom_vline(xintercept = 0, color ="#9400d3") +
  geom_vline(xintercept = 10, color ="#9400d3") +
  theme(axis.text.y=element_blank())
ggsave("../figures/prior_predict.png", prior_predict, width = 5, height = 3, units = "in")
```

```{r}
# plotting the priors
p_intercept_sim <-
  tibble(x = seq(from = -10, to = 10, by = 0.1)) %>%
  
  ggplot(aes(x = x, y = dnorm(x, mean = 8, sd = 2))) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  labs(title = "Intercept", x="", y="") +
  theme(text = element_text(size=14))
p_intercept_sim

p_b_sim <-
  tibble(x = seq(from = -10, to = 10, by = 0.1)) %>%
  
  ggplot(aes(x = x, y = dnorm(x, mean = 0, sd = 1))) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  labs(title = "Population level effects", x="", y="") +
  theme(text = element_text(size=14))
p_b_sim

p_sd_sim <-
  tibble(x = seq(from = -10, to = 10, by = 0.1)) %>%
  
  ggplot(aes(x = x, y = dcauchy(x, location = 0, scale = 1))) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  labs(title = "Group level effects", x="", y="") +
  theme(text = element_text(size=14))
p_sd_sim

p_sigma_sim <-
  tibble(x = seq(from = -10, to = 10, by = 0.1)) %>%
  
  ggplot(aes(x = x, y = dexp(x, rate = 0.5))) +
  geom_line(color='#004c99') +
  scale_x_continuous(breaks = seq(from = -10, to = 10, by = 2)) +
  labs(title = "Sigma", x="", y="") +
  theme(text = element_text(size=14))
p_sigma_sim

plot_priors <- grid.arrange(p_intercept_sim, p_b_sim, p_sd_sim, p_sigma_sim)
ggsave("../figures/plot_priors.png", plot_priors, width = 7, height = 4, units = "in")
plot_priors

test <- ggarrange(p_intercept_sim, p_b_sim, p_sd_sim, p_sigma_sim)
annotate_figure(test,
                bottom = "x",
                left = "Density")
```

```{r}
m3.1 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(7, 2), class = Intercept),
                     prior(normal(0, 0.8), class = b),
                     prior(cauchy(0, 1), class = sd),
                     prior(exponential(0.5), class = sigma)),
                     file = "../results/prior_m31")

m3.2 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(7, 2), class = Intercept),
                     prior(normal(0, 0.8), class = b),
                     prior(cauchy(0, 5), class = sd),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_m32")

m3.3 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(7, 2), class = Intercept),
                     prior(normal(0, 0.8), class = b),
                     prior(cauchy(0, 10), class = sd),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_m33")

m3.4 <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(normal(7, 2), class = Intercept),
                     prior(normal(0, 0.8), class = b),
                     prior(cauchy(0, 50), class = sd),
                     prior(exponential(0.5), class = sigma)),
           file = "../results/prior_m34")
```

```{r}
m3.1_predict <- posterior_epred(object = m3.1, nsamples = 100)
m3.1_long <- gather(data.frame(m3.1_predict))
m3.1_long
prior31 <- ggplot(m3.1_long, aes(x=value, group=key)) +
  geom_histogram(alpha=.2, bins = 200, fill = "#011f4b") +
  xlim(-50,50) +
  xlab("Predicted life satisfaction") + 
  ylab("Density") +
  labs(title = "prior(cauchy(0,1), class = sd") +
  theme(axis.text.y = element_blank())
```
```{r}
m3.2_predict <- posterior_epred(object = m3.2, nsamples = 100)
m3.2_long <- gather(data.frame(m3.2_predict))
m3.2_long
prior32 <- ggplot(m3.2_long, aes(x=value, group=key)) +
  geom_histogram(alpha=.2, bins = 200, fill = "#011f4b") +
  xlim(-50,50) +
  xlab("Predicted life satisfaction") + 
  ylab("Density") +
  labs(title = "prior(cauchy(0,5), class = sd") +
  theme(axis.text.y = element_blank())
```
```{r}
m3.3_predict <- posterior_epred(object = m3.3, nsamples = 100)
m3.3_long <- gather(data.frame(m3.3_predict))
m3.3_long
prior33 <- ggplot(m3.3_long, aes(x=value, group=key)) +
  geom_histogram(alpha=.2, bins = 200, fill = "#011f4b") +
  xlim(-50,50) +
  xlab("Predicted life satisfaction") + 
  ylab("Density") +
  labs(title = "prior(cauchy(0,10), class = sd") +
  theme(axis.text.y = element_blank())
```
```{r}
m3.4_predict <- posterior_epred(object = m3.3, nsamples = 100)
m3.4_long <- gather(data.frame(m3.3_predict))
m3.4_long
prior34 <- ggplot(m3.4_long, aes(x=value, group=key)) +
  geom_histogram(alpha=.2, bins = 200, fill = "#011f4b") +
  xlim(-50,50) +
  xlab("Predicted life satisfaction") + 
  ylab("Density") +
  labs(title = "prior(cauchy(0,50), class = sd") +
  theme(axis.text.y = element_blank())
```

```{r}
grid.arrange(prior31, prior32, prior33, prior34, ncol = 2)
```


```{r}
# Let's take prior21
# The likelihood for the model is h ~ N(\mu,\sigma)
# Our \mu prior will be \mu ~ MODEL
# Our prior for \sigma will be \sigma ~ Cauchy(0,1)

# Here is the shape for the prior \mu 

                           
intercept <- ggplot(data = tibble(x = seq(from = 0, to = 15, by = .1)),
       aes(x=x, y = dnorm(x, mean = 8, sd = 2))) +
  geom_line(colour = "#6497b1", size = 1.5) +
  ylab("density")

b <- ggplot(data = tibble(x = seq(from = -5, to = 5, by = .1)),
       aes(x=x, y = dnorm(x, mean = 0, sd = 0.75))) +
  geom_line(colour = "#6497b1", size = 1.5) +
  ylab("density")

sd <- ggplot(data = tibble(x = seq(from = -5, to = 5, by = .1)),
       aes(x=x, y = dcauchy(x, location = 0, scale = 1))) +
  geom_line(colour = "#6497b1", size = 1.5) +
  ylab("density")

sigma <- ggplot(data = tibble(x = seq(from = -5, to = 5, by = .1)),
       aes(x=x, y = dexp(x, rate = 1))) +
  geom_line(colour = "#6497b1", size = 1.5) +
  ylab("density")

chosen_priors <- plot_grid(intercept, b, sd, sigma, labels = c("Intercept", "b", "sd", "sigma"))
ggsave("../figures/chosen_priors.png", chosen_priors, width = 7, height = 4, units = "in")
```

```{r}
m2.1test <- brm( lsat ~ treated + welle  + tXw2 + tXw3 + tXw4 + tXw5 + female + migback + hob.mus.making + (1 + treated + tXw2 + tXw3 + tXw4 + tXw5 | school) + (1 | pid),
           sample_prior = "only", data = d,
           prior = c(prior(rnorm(1e4, 7, 2), class = Intercept),
                     prior(rnorm(1e4, 0, 0.75), class = b),
                     prior(rcauchy(1e4, 0, 1), class = sd),
                     prior(rexp(1e4, 0.5), class = sigma)))

m2.1test <- brm(lsat ~ treated + welle  + tXw2 + female +
                   (1 + treated + tXw2 | school) +
                   (1 | pid),
           sample_prior = "only", data = subset(d, welle == 1 | welle == 2),
           prior = c(prior(normal(7, 2), class = Intercept),
                     prior(normal(0, 0.75), class = b),
                     prior(cauchy(0, 1), class = sd),
                     prior(exponential(0.5), class = sigma)))
lsat 

sample_intercept <- rnorm(1e4, 7, 2)
prior_intercept <- mean(sample_intercept)
sample_intercept
prior_intercept
```
